Integrate Samplers into Concentrator buckets
Rather than instantiating samplers in the writer and trying to align
writer buffers on concentrator buckets, we know directly pass a
`concentratorBuffer` object to the Writer which contains the stats and
a sampler where the spans have been added.
Avoids blips and late spans to be added to the wrong buffer.