Replace service by layer and add fixtures package
Using the "layer" terminology instead of "service", layers are
basically strings that can be a concatenation of several entities
separated by dots '.'.
The first part of the layer (before the first dot) is considered as the
"app" and all the following levels are sub-layers of this app.
(NOTE: the team decided it will probably evolve into a dedicated "App"
field in the span).

The other big change here is adding a new fixtures/ folder which
defines its own package. It serves the purpose of generating fake
spans for tests or fuzz mode.