Allow configuring concentrator OldestSpanCutoff
This was hardcoded at 5 before.
This controls basically how far back in time we allow a span to be added
to an older bucket, this has 2 consequences: spans older than oldest
span cutoff are dropped, and we wait at least oldest span cutoff to
flush any bucket to make sure no spans will be added to them.

Our bucket was 10s large, meaning we would drop as late spans half of
them.