[config] allow using api_key array
The dd-agent config that we rely on to get the API key of the user
will now allow a comma-separated list of values to enable sending
the data to more than one account.
We allow this new behavior by parsing the config accordingly and we
handle the duplication of data in the agent writer.

Note that when there is only one API key, nothing changes.