[profile] unpacking "_resource" as resource in tags
In many implementations, resource is sent in the resource field and
as a tag, since agent can decide to treat them differently. Typically,
resource would be quantized, but not the value in the tagset. With
this patch, a client implemetation can just send "_resource" instead
of the complete resource value, to save bandwith.