add a /status route
This route returns the current status of the agent. It currently only
contains build information, and a `running` flag, but now we have
something that can be extended in the future.

This feature is added because it provides a simple way to check the
status of an agent.

Storing the status structure in the agent instead of the receiver and
including concurrent access protection might sound excessive, but it
will be useful to update its content from other parts of the agent.