add workaround for LimitedReader to support msgp
It's perfectly normal for a reader to return n>0 and EOF:

    Callers should always process the n > 0 bytes returned before
    considering the error err. Doing so correctly handles I/O errors
    that happen after reading some bytes and also both of the allowed
    EOF behaviors.
        -- https://golang.org/pkg/io/#Reader

Infortunately, github.com/tinylib/msgp does not handle this case, and
return immediately if err != nil, without checking whether data were
read.

This workaround set the error to nil if there are data available. The
next read will return n=0 and EOF.