weight spans using their sample rate
The root span of a trace can contain a _sample_rate value if the client
sampled spans. When this is the case, we scale stats (hit and error
counts, and durations) by a factor, or weight, equal to 1.0 /
_sample_rate.

This also means that we have to use floating point values to accumulate
values in StatsRawBucket.