[log] using receiver limited logger instead of plain log debug
When skipping messages, a debug message is of no use in production mode,
as we get no clue on why traces where dropped. Especially when it's for
an obscure reason such as clock skew. This patch uses the receiver logger
object which has a limiter preventing log to be filled with the same error
repeating many times, while still displaying a few of them.