config: add support for SOCKS5 and deprecate proxy.http(s)
This change adds support for SOCKS5 proxies (which was in theory already
present via the already existing proxy settings and their mechanics, but
confusingly implemented) using the `proxy.url` setting.

It also deprecates the unnecessary `proxy.http` and `proxy.https`
settings in the example files and marks them as such. The scheme can
already be deduced from the URL.