internal/api: move "trace dropped" log to level DEBUG
This change moves the log entry that notifies the user that his trace
was dropped (due to the agent being too slow) to Debug, as it has turned
out to become quite spammy in high traffic environments. It is also not
of much value to the user. Dropping is already counted as a metric and
will be visible in the app.