internal/statsd: rename to metrics
This change renames the statsd package to metrics and adds some helper
methods to simplify the calls.

This is preliminary work for splitting the `agent` package into other
packages, one of which will be called `stats`. `metrics` is a better
name because it matches our product terminology and will reduce the
chance of confusing its functions with that of APM-computed stats.